게시판 예외추가(1)

+++ src/main/java/site/ch00kh/domain/post/application/PostService.java
@@ -2,6 +2,7 @@ package site.ch00kh.domain.post.application;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.coyote.BadRequestException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -10,9 +11,11 @@ import org.springframework.transaction.annotation.Transactional;
 import site.ch00kh.domain.account.dto.AccountDetails;
 import site.ch00kh.domain.post.dao.Post;
 import site.ch00kh.domain.post.dao.PostRepository;
+import site.ch00kh.domain.post.dto.PostDeleteRequestDto;
 import site.ch00kh.domain.post.dto.PostPageRequestDto;
 import site.ch00kh.domain.post.dto.PostWriteRequestDto;
 import site.ch00kh.domain.post.dto.PostWriteResponseDto;
+import site.ch00kh.domain.post.exception.PostNotFoundException;
 
 @Slf4j
 @Service
@@ -26,15 +29,13 @@ public class PostService {
     @Transactional(readOnly = true)
     public PostWriteResponseDto read(Long id) {
 
-        Post post = postRepository.findById(id).orElseThrow();
+        Post post = postRepository.findByIdAndIsDeleted(id, "N").orElseThrow(PostNotFoundException::new);
         return PostWriteResponseDto.from(post);
     }
 
     @Transactional(readOnly = true)
     public Page<PostWriteResponseDto> readAll(PostPageRequestDto postPageRequestDto) {
 
-        log.info("readAll");
-        log.info("{}", postPageRequestDto);
         Pageable pageable = postPageRequestDto.toPageable();
         Page<Post> postPage = postRepository.findByIsDeleted("N", pageable);
 
@@ -50,30 +51,40 @@ public class PostService {
         return PostWriteResponseDto.from(savedPost);
     }
 
-    public PostWriteResponseDto modify(Long id, PostWriteRequestDto requestDto, AccountDetails accountDetails) {
+    public PostWriteResponseDto modify(Long id, PostWriteRequestDto requestDto, AccountDetails accountDetails) throws BadRequestException {
 
-        Post validPost = isValid(id, requestDto, accountDetails);
+        String password = requestDto.getPassword();
+        String username = accountDetails.getUsername();
+
+        Post validPost = isValid(id, password, username);
         Post modifiedPost = validPost.modify(requestDto);
 
         return PostWriteResponseDto.from(modifiedPost);
     }
 
-    public void delete(Long id, PostWriteRequestDto requestDto, AccountDetails accountDetails) {
+    public void delete(Long id, PostDeleteRequestDto requestDto, AccountDetails accountDetails) throws BadRequestException {
 
-        Post validPost = isValid(id, requestDto, accountDetails);
+        String password = requestDto.getPassword();
+        String username = accountDetails.getUsername();
+
+        Post validPost = isValid(id, password, username);
+
+        if (validPost.getIsDeleted().equalsIgnoreCase("Y")) {
+            throw new PostNotFoundException();
+        }
 
         validPost.delete();
     }
 
-    private Post isValid(Long id, PostWriteRequestDto requestDto, AccountDetails accountDetails) {
+    private Post isValid(Long id, String password, String username) throws BadRequestException {
         Post post = postRepository.findById(id).orElseThrow(RuntimeException::new);
 
-        if (!bCryptPasswordEncoder.matches(requestDto.getPassword(), post.getPassword())) {
-            throw new RuntimeException();
+        if (!bCryptPasswordEncoder.matches(password, post.getPassword())) {
+            throw new BadRequestException();
         }
 
-        if (!post.getWriter().equals(accountDetails.getUsername())) {
-            throw new RuntimeException();
+        if (!post.getWriter().equals(username)) {
+            throw new BadRequestException();
         }
 
         return post;
